Refactored paste links into a plugin
