+++ src/app/components/Composer/plugins/index.js
@@ -1,10 +1,13 @@
 // components
 import { MarkHotkey } from "./mark-hotkey"
+import { Linkify } from "./linkify"
 import AutoReplace from "slate-auto-replace"
 import EditBlockquote from "slate-edit-blockquote"
 
 // export
 export const plugins = [
+	// tools to insert links into text
+	Linkify,
 	// hot keys
   MarkHotkey({ 	key: "b", 			type: "bold" }),
   MarkHotkey({ 	key: "i", 			type: "italic" }),
